fixed hw 10
