+++ src/by/it/academy/hw10/task2/Test.java
@@ -1,0 +1,28 @@
+package by.it.academy.hw10.task2;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Scanner;
+
+public class Test {
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        System.out.println("Enter text");
+        String text = sc.nextLine();
+        Map<String, Integer> map = new HashMap<>();
+        String[] words = text.split("\\s+");
+        for (String word : words) {
+            word = word.replaceAll("[^a-zA-ZaА-Я0-9]", "");
+            if (word.length() > 0) {
+                if (map.containsKey(word)) {
+                    map.put(word, map.get(word) + 1);
+                } else {
+                    map.put(word, 1);
+                }
+            }
+        }
+        for (Map.Entry<String, Integer> entry : map.entrySet()) {
+            System.out.println(entry.getKey() + " - " + entry.getValue());
+        }
+    }
+}
