+++ src/by/it/academy/hw10/task3/Money.java
@@ -1,22 +1,8 @@
-package by.it.academy.hw10;
-
+package by.it.academy.hw10.task3;
 
 import java.util.Stack;
 
 public class Money {
-    private final int number;
-
-    public Money(int number) {
-        this.number = number;
-    }
-
-    @Override
-    public String toString() {
-        return "Cards{" +
-                "number=" + number +
-                '}';
-    }
-
     public static Integer max(Stack<Integer> stack) {
         if (stack.isEmpty()) {
             return Integer.MIN_VALUE;
